Refactor into number utils.
