+++ src/example-types/number.js
@@ -1,69 +1,9 @@
 const _ = require('lodash/fp')
-/*
- 1. An empty value as the upper boundary represents infinity.
- 2. An empty value as the lower boundary represents negative infinity.
- 3. Zero has to be respected as a boundary value.
-*/
-let boundaryFilter = value => {
-  if (_.isString(value) && _.isEmpty(value)) value = NaN
-  return _.isNaN(_.toNumber(value)) ? null : _.toNumber(value)
-}
-
-let rangeFilter = (field, min, max) => ({
-  range: {
-    [field]: _.pickBy(_.isNumber, {
-      gte: boundaryFilter(min),
-      lte: boundaryFilter(max),
-    }),
-  },
-})
-
-let searchResults = async (search, field, min, max, percentileInterval) => {
-  let statisticalResult = await search({
-    aggs: {
-      range_filter: {
-        filter: rangeFilter(field, min, max),
-        aggs: {
-          statistical: {
-            stats: {
-              field,
-              missing: 0,
-            },
-          },
-          all_percentiles: {
-            percentiles: {
-              field,
-              percents: [0, percentileInterval, 100 - percentileInterval, 100],
-            },
-          },
-        },
-      },
-    },
-  })
-
-  let percentiles = _.flow(_.mapKeys(Number), mappedResult => ({
-    rangeMin: mappedResult[0],
-    rangeMax: mappedResult[100],
-    intervalMin: mappedResult[percentileInterval],
-    intervalMax: mappedResult[100 - percentileInterval],
-  }))(
-    _.get('aggregations.range_filter.all_percentiles.values', statisticalResult)
-  )
-
-  let statistical = _.get(
-    'aggregations.range_filter.statistical',
-    statisticalResult
-  )
-
-  return {
-    statistical,
-    percentiles,
-  }
-}
+const util = require('./numberUtil')
 
 module.exports = {
   hasValue: context => !_.isNil(context.min) || !_.isNil(context.max),
-  filter: ({ field, min, max }) => rangeFilter(field, min, max),
+  filter: ({ field, min, max }) => util.rangeFilter(field, min, max),
   async result(
     {
       field,
@@ -86,13 +26,14 @@ module.exports = {
       let iterationCount = 1
 
       while (hasMaxOutlier || hasMaxOutlier || iterationCount <= maxIteration) {
-        results = await searchResults(
+        results = await util.getStatisticalResults(
           search,
           field,
           minValue,
           maxValue,
           percentileInterval
         )
+        
         let { statistical, percentiles } = results
         let rangeMin = _.get('min', statistical)
         let rangeMax = _.get('max', statistical)
@@ -121,7 +62,7 @@ module.exports = {
         },
       })
     } else {
-      results = await searchResults(search, field, min, max, percentileInterval)
+      results = await util.getStatisticalResults(search, field, min, max, percentileInterval)
     }
 
     return results
