+++ src/example-types/numberRangeHistogram.js
@@ -1,70 +1,16 @@
 const _ = require('lodash/fp')
-
-/*
- 1. An empty value as the upper boundary represents infinity.
- 2. An empty value as the lower boundary represents negative infinity.
- 3. Zero has to be respected as a boundary value.
-*/
-let boundaryFilter = value => {
-  if (_.isString(value) && _.isEmpty(value)) value = NaN
-  return _.isNaN(_.toNumber(value)) ? null : _.toNumber(value)
-}
-
-let rangeFilter = (field, min, max) => ({
-  range: {
-    [field]: _.pickBy(_.isNumber, {
-      gte: boundaryFilter(min),
-      lte: boundaryFilter(max),
-    }),
-  },
-})
+const util = require('./numberUtil')
 
 module.exports = {
   hasValue: context => !_.isNil(context.min) || !_.isNil(context.max),
-  filter: ({ field, min, max }) => rangeFilter(field, min, max),
+  filter: ({ field, min, max }) => util.rangeFilter(field, min, max),
   async result({ field, min, max, percentileInterval = 1 }, search) {
-    let statisticalResult = await search({
-      aggs: {
-        range_filter: {
-          filter: rangeFilter(field, min, max),
-          aggs: {
-            statistical: {
-              stats: {
-                field,
-                missing: 0,
-              },
-            },
-            all_percentiles: {
-              percentiles: {
-                field,
-                percents: [
-                  0,
-                  percentileInterval,
-                  100 - percentileInterval,
-                  100,
-                ],
-              },
-            },
-          },
-        },
-      },
-    })
-
-    let percentiles = _.flow(_.mapKeys(Number), mappedResult => ({
-      rangeMin: mappedResult[0],
-      rangeMax: mappedResult[100],
-      intervalMin: mappedResult[percentileInterval],
-      intervalMax: mappedResult[100 - percentileInterval],
-    }))(
-      _.get(
-        'aggregations.range_filter.all_percentiles.values',
-        statisticalResult
-      )
-    )
-
-    let statistical = _.get(
-      'aggregations.range_filter.statistical',
-      statisticalResult
+    let { statistical, percentiles } = await util.getStatisticalResults(
+      search,
+      field,
+      minValue,
+      maxValue,
+      percentileInterval
     )
 
     let interval =
@@ -75,7 +21,7 @@ module.exports = {
       let histogramResult = await search({
         aggs: {
           range_filter: {
-            filter: rangeFilter(field, min, max),
+            filter: util.rangeFilter(field, min, max),
             aggs: {
               values: {
                 histogram: {
